requirements.txt
# Dependencies for University Timetable Automation
# Install with: pip install -r requirements.txt

requests>=2.28.0
beautifulsoup4>=4.11.0
python-pptx>=0.6.21
Pillow>=9.0.0
pypdfium2>=4.0.0
pyinstaller>=5.0
